consolidated processing of antibody annotation files
